pyproject.toml
[project]
name = "euclid-macro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.8.18",
    "fredapi>=0.5.2",
    "myeia>=0.4.8",
    "numpy>=2.3.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "pyreadr>=0.5.3",
    "python-dotenv>=1.1.1",
    "pyxlsb>=1.0.10",
    "requests>=2.32.4",
    "scipy>=1.16.0",
    "selenium>=4.34.2",
    "selenium-stealth>=1.0.6",
    "webdriver-manager>=4.0.2",
    "yfinance>=0.2.65",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
]
